requirements.txt
attrs==23.1.0
aws-cdk-lib==2.104.0
aws-cdk.asset-awscli-v1==2.2.201
aws-cdk.asset-kubectl-v20==2.1.2
aws-cdk.asset-node-proxy-agent-v6==2.0.1
cattrs==23.1.2
constructs==10.3.0
importlib-resources==6.1.0
jsii==1.91.0
publication==0.0.3
python-dateutil==2.8.2
six==1.16.0
typeguard==2.13.3
typing_extensions==4.8.0
